修改

+++ contracts/PracticeContract.sol
@@ -16,7 +16,7 @@ contract PracticeContract {
 
     function batchProcess(uint256[] calldata inputs) external {
         require(msg.sender == ADMIN, "not admin");
-        uint len = inputs.length;
+        uint256 len = inputs.length;
         uint256 result;
         for (uint i= 0; i < len; i++) {
             result = inputs[i] * MULTIPLIER;
@@ -27,7 +27,7 @@ contract PracticeContract {
     function getSum() external view returns (uint256) {
         require(msg.sender == ADMIN, "not admin");
         uint256 sum = 0;
-        uint len = numbers.length;
+        uint256 len = numbers.length;
         for (uint i= 0; i < len; i++) {
             sum += numbers[i];
         }
